Setup core event handler
