+++ trin-core/src/portalnet/mod.rs
@@ -5,6 +5,7 @@ use ssz::DecodeError;
 use uint::construct_uint;
 
 pub mod discovery;
+pub mod events;
 pub mod overlay;
 pub mod protocol;
 pub mod types;
